First attempt (lets hope everything is correct)
New King Safety works with more linear terms (still need special tuning) and works like this:
We collect the amount of attacked squares into the enemy king zone which are not defended from all pieces and multiply those with the PIECE_ATTACK_BASE_VALUE and sum it up. The same is done with all safe checks the pieces can give to the enemy king. The sum of those two values is the ATTACK_BASE_VALUE. For each attacking Piece, its ATTACK_FORCE will be summed up to an Attack force accumulator. the Attack base value will be adjusted by a quadratic term depending on the attack force, so that the more pieces are attacking the higher the evaluation gets, and if only one piece is attacking the evaluation will be low.
BENCH=11323116